renamed CrashMap to PointMap
